Move hooks to src
