+++ src/hooks/useFavorites.test.ts
